requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
idna==3.10
python-decouple==3.8
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
